# Conflicts:
#	contrib/opentimelineio_contrib/adapters/ffmpeg_burnins.py

+++ tests/utils.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2017 Pixar Animation Studios
+# Copyright Contributors to the OpenTimelineIO project
 #
 # Licensed under the Apache License, Version 2.0 (the "Apache License")
 # with the following modification; you may not use this file except in
@@ -53,4 +53,6 @@ def create_manifest():
 def remove_manifest(manifest):
     """Remove the manifest source files."""
     for file_path in manifest.source_files:
-        os.remove(file_path)
+        # don't accidentally blow away python
+        if not file_path.endswith('.py'):
+            os.remove(file_path)
